+++ lib/Expression/Func/TimeConvertFunction.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace PhpBench\Expression\Func;
+
+use PhpBench\Expression\Ast\FloatNode;
+use PhpBench\Expression\Ast\NumberValue;
+use PhpBench\Expression\Ast\PhpValue;
+use PhpBench\Util\TimeUnit;
+
+final class TimeConvertFunction
+{
+    public function __construct(private TimeUnit $timeUnit)
+    {
+    }
+
+    public function __invoke(NumberValue $value, PhpValue $from, PhpValue $to): PhpValue
+    {
+        $from = $from->value();
+        $to = $to->value();
+
+        $result = $this->timeUnit->convertTo($value->value(), (string)$from, (string)$to);
+
+        return new FloatNode($result);
+    }
+}